Statistics Data Digester
Contains digestData, which takes a first and last included season and returns a tuple of two dictionaries, both organized by year. The first dictionary contains a DataTable with statistics for every team in the given year, the second contains a glossary of each column title in the first dictionary's DataTable for each year.